[skip ci] :recycle: api: move csv header generation to src
