+++ api/src/certification/session/application/session-mass-import-controller.js
@@ -1,6 +1,6 @@
 import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import { usecases } from '../../../../src/certification/shared/domain/usecases/index.js';
-import { getHeaders } from '../../../../lib/infrastructure/files/sessions-import.js';
+import { getHeaders } from '../infrastructure/files/sessions-import.js';
 
 import * as csvHelpers from '../../../../lib/application/certification-centers/csvHelpers.js';
 import * as csvSerializer from '../../../../lib/infrastructure/serializers/csv/csv-serializer.js';
